refactor: epoch settings message moved to message service
