+++ mithril-aggregator/src/http_server/routes/epoch_routes.rs
@@ -1,13 +1,5 @@
-use std::collections::BTreeSet;
 use warp::Filter;
 
-use mithril_common::{
-    entities::SignedEntityTypeDiscriminants,
-    messages::{EpochSettingsMessage, SignerMessagePart},
-    StdResult,
-};
-
-use crate::dependency_injection::EpochServiceWrapper;
 use crate::http_server::routes::middlewares;
 use crate::http_server::routes::router::RouterState;
 
@@ -24,70 +16,34 @@ fn epoch_settings(
     warp::path!("epoch-settings")
         .and(warp::get())
         .and(middlewares::with_logger(router_state))
-        .and(middlewares::with_epoch_service(router_state))
+        .and(middlewares::with_http_message_service(router_state))
         .and(middlewares::extract_config(router_state, |config| {
             config.allowed_discriminants.clone()
         }))
         .and_then(handlers::epoch_settings)
 }
 
-async fn get_epoch_settings_message(
-    epoch_service: EpochServiceWrapper,
-    allowed_discriminants: BTreeSet<SignedEntityTypeDiscriminants>,
-) -> StdResult<EpochSettingsMessage> {
-    let epoch_service = epoch_service.read().await;
-
-    let epoch = epoch_service.epoch_of_current_data()?;
-    let signer_registration_protocol_parameters = epoch_service
-        .signer_registration_protocol_parameters()?
-        .clone();
-    let current_signers = epoch_service.current_signers()?;
-    let next_signers = epoch_service.next_signers()?;
-
-    let cardano_transactions_discriminant =
-        allowed_discriminants.get(&SignedEntityTypeDiscriminants::CardanoTransactions);
-
-    let cardano_transactions_signing_config = cardano_transactions_discriminant
-        .map(|_| epoch_service.current_cardano_transactions_signing_config())
-        .transpose()?
-        .cloned();
-    let next_cardano_transactions_signing_config = cardano_transactions_discriminant
-        .map(|_| epoch_service.next_cardano_transactions_signing_config())
-        .transpose()?
-        .cloned();
-
-    let epoch_settings_message = EpochSettingsMessage {
-        epoch,
-        signer_registration_protocol_parameters,
-        current_signers: SignerMessagePart::from_signers(current_signers.to_vec()),
-        next_signers: SignerMessagePart::from_signers(next_signers.to_vec()),
-        cardano_transactions_signing_config,
-        next_cardano_transactions_signing_config,
-    };
-
-    Ok(epoch_settings_message)
-}
-
 mod handlers {
     use slog::{warn, Logger};
     use std::collections::BTreeSet;
     use std::convert::Infallible;
+    use std::sync::Arc;
     use warp::http::StatusCode;
 
     use mithril_common::entities::SignedEntityTypeDiscriminants;
 
-    use crate::dependency_injection::EpochServiceWrapper;
-    use crate::http_server::routes::epoch_routes::get_epoch_settings_message;
     use crate::http_server::routes::reply;
+    use crate::services::MessageService;
 
     /// Epoch Settings
     pub async fn epoch_settings(
         logger: Logger,
-        epoch_service: EpochServiceWrapper,
+        http_message_service: Arc<dyn MessageService>,
         allowed_discriminants: BTreeSet<SignedEntityTypeDiscriminants>,
     ) -> Result<impl warp::Reply, Infallible> {
-        let epoch_settings_message =
-            get_epoch_settings_message(epoch_service, allowed_discriminants).await;
+        let epoch_settings_message = http_message_service
+            .get_epoch_settings_message(allowed_discriminants)
+            .await;
 
         match epoch_settings_message {
             Ok(message) => Ok(reply::json(&message, StatusCode::OK)),
@@ -101,26 +57,17 @@ mod handlers {
 
 #[cfg(test)]
 mod tests {
+    use anyhow::anyhow;
     use serde_json::Value::Null;
-    use std::collections::BTreeSet;
     use std::sync::Arc;
-    use tokio::sync::RwLock;
     use warp::{
         http::{Method, StatusCode},
         test::request,
     };
 
-    use mithril_common::{
-        entities::{
-            BlockNumber, CardanoTransactionsSigningConfig, Epoch, ProtocolParameters,
-            SignedEntityTypeDiscriminants,
-        },
-        test_utils::{apispec::APISpec, fake_data, MithrilFixtureBuilder},
-    };
+    use mithril_common::{messages::EpochSettingsMessage, test_utils::apispec::APISpec};
 
-    use crate::entities::AggregatorEpochSettings;
-    use crate::services::FakeEpochService;
-    use crate::{initialize_dependencies, services::FakeEpochServiceBuilder};
+    use crate::{initialize_dependencies, services::MockMessageService};
 
     use super::*;
 
@@ -136,126 +83,16 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn get_epoch_settings_message_with_cardano_transactions_enabled() {
-        let fixture = MithrilFixtureBuilder::default().with_signers(3).build();
-        let epoch_service = FakeEpochService::from_fixture(Epoch(4), &fixture);
-        let epoch_service = Arc::new(RwLock::new(epoch_service));
-        let allowed_discriminants =
-            BTreeSet::from([SignedEntityTypeDiscriminants::CardanoTransactions]);
-
-        let message = get_epoch_settings_message(epoch_service, allowed_discriminants)
-            .await
-            .unwrap();
-
-        assert!(message.cardano_transactions_signing_config.is_some());
-        assert!(message.next_cardano_transactions_signing_config.is_some(),);
-    }
-
-    #[tokio::test]
-    async fn get_epoch_settings_message_with_cardano_transactions_not_enabled() {
-        let fixture = MithrilFixtureBuilder::default().with_signers(3).build();
-        let epoch_service = FakeEpochService::from_fixture(Epoch(4), &fixture);
-        let epoch_service = Arc::new(RwLock::new(epoch_service));
-
-        let allowed_discriminants = BTreeSet::new();
-        let message = get_epoch_settings_message(epoch_service, allowed_discriminants)
-            .await
-            .unwrap();
-
-        assert_eq!(message.cardano_transactions_signing_config, None);
-        assert_eq!(message.next_cardano_transactions_signing_config, None);
-    }
-
-    #[tokio::test]
-    async fn get_epoch_settings_message_retrieves_protocol_parameters_from_epoch_service() {
-        let current_epoch_settings = AggregatorEpochSettings {
-            protocol_parameters: ProtocolParameters::new(101, 10, 0.5),
-            ..AggregatorEpochSettings::dummy()
-        };
-        let next_epoch_settings = AggregatorEpochSettings {
-            protocol_parameters: ProtocolParameters::new(102, 20, 0.5),
-            ..AggregatorEpochSettings::dummy()
-        };
-        let signer_registration_epoch_settings = AggregatorEpochSettings {
-            protocol_parameters: ProtocolParameters::new(103, 30, 0.5),
-            ..AggregatorEpochSettings::dummy()
-        };
-
-        let epoch_service = FakeEpochServiceBuilder {
-            current_epoch_settings,
-            next_epoch_settings: next_epoch_settings.clone(),
-            signer_registration_epoch_settings: signer_registration_epoch_settings.clone(),
-            current_signers_with_stake: fake_data::signers_with_stakes(5),
-            next_signers_with_stake: fake_data::signers_with_stakes(3),
-            ..FakeEpochServiceBuilder::dummy(Epoch(1))
-        }
-        .build();
-
-        let message = get_epoch_settings_message(
-            Arc::new(RwLock::new(epoch_service)),
-            SignedEntityTypeDiscriminants::all(),
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(
-            message.signer_registration_protocol_parameters,
-            signer_registration_epoch_settings.protocol_parameters
-        );
-    }
-
-    #[tokio::test]
-    async fn get_epoch_settings_message_retrieves_signing_configuration_from_epoch_service() {
-        let current_epoch_settings = AggregatorEpochSettings {
-            cardano_transactions_signing_config: CardanoTransactionsSigningConfig::new(
-                BlockNumber(100),
-                BlockNumber(15),
-            ),
-            ..AggregatorEpochSettings::dummy()
-        };
-        let next_epoch_settings = AggregatorEpochSettings {
-            cardano_transactions_signing_config: CardanoTransactionsSigningConfig::new(
-                BlockNumber(200),
-                BlockNumber(15),
-            ),
-            ..AggregatorEpochSettings::dummy()
-        };
-
-        let epoch_service = FakeEpochServiceBuilder {
-            current_epoch_settings: current_epoch_settings.clone(),
-            next_epoch_settings: next_epoch_settings.clone(),
-            signer_registration_epoch_settings: AggregatorEpochSettings::dummy(),
-            current_signers_with_stake: fake_data::signers_with_stakes(5),
-            next_signers_with_stake: fake_data::signers_with_stakes(3),
-            ..FakeEpochServiceBuilder::dummy(Epoch(1))
-        }
-        .build();
-
-        let message = get_epoch_settings_message(
-            Arc::new(RwLock::new(epoch_service)),
-            SignedEntityTypeDiscriminants::all(),
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(
-            message.cardano_transactions_signing_config,
-            Some(current_epoch_settings.cardano_transactions_signing_config),
-        );
-        assert_eq!(
-            message.next_cardano_transactions_signing_config,
-            Some(next_epoch_settings.cardano_transactions_signing_config),
-        );
-    }
-
-    #[tokio::test]
     async fn test_epoch_settings_get_ok() {
         let method = Method::GET.as_str();
         let path = "/epoch-settings";
         let mut dependency_manager = initialize_dependencies().await;
-        let fixture = MithrilFixtureBuilder::default().with_signers(5).build();
-        let epoch_service = FakeEpochService::from_fixture(Epoch(5), &fixture);
-        dependency_manager.epoch_service = Arc::new(RwLock::new(epoch_service));
+        let mut mock_http_message_service = MockMessageService::new();
+        mock_http_message_service
+            .expect_get_epoch_settings_message()
+            .return_once(|_| Ok(EpochSettingsMessage::dummy()))
+            .once();
+        dependency_manager.message_service = Arc::new(mock_http_message_service);
 
         let response = request()
             .method(method)
@@ -281,7 +118,13 @@ mod tests {
     async fn test_epoch_settings_get_ko_500() {
         let method = Method::GET.as_str();
         let path = "/epoch-settings";
-        let dependency_manager = initialize_dependencies().await;
+        let mut dependency_manager = initialize_dependencies().await;
+        let mut mock_http_message_service = MockMessageService::new();
+        mock_http_message_service
+            .expect_get_epoch_settings_message()
+            .return_once(|_| Err(anyhow!("an error")))
+            .once();
+        dependency_manager.message_service = Arc::new(mock_http_message_service);
 
         let response = request()
             .method(method)
